#268 LH2: Copy new implementation over to the correct file  lh2.c/.h
