--- projects/01bsp_lighthouse_4/README.md
@@ -1,4 +1,0 @@
-# Lighthouse v2 board support package usage example
-
-
-## Not Implemented
