+++ projects/projects-bsp-drv.emProject
@@ -117,45 +117,6 @@
       <file file_name="$(SeggerThumbStartup)" />
     </folder>
   </project>
-  <project Name="01bsp_lighthouse_4">
-    <configuration
-      Name="Common"
-      project_dependencies="00bsp_dotbot_board(bsp);00bsp_dotbot_lh2_4(bsp)"
-      project_directory="01bsp_lighthouse_4"
-      project_type="Executable" />
-    <folder Name="Device Files">
-      <file file_name="$(DeviceCommonHeaderFile)" />
-      <file file_name="$(DeviceHeaderFile)" />
-      <file file_name="$(DeviceSystemFile)">
-        <configuration
-          Name="Common"
-          default_code_section=".init"
-          default_const_section=".init_rodata" />
-      </file>
-    </folder>
-    <folder Name="Script Files">
-      <file file_name="$(DeviceLinkerScript)">
-        <configuration Name="Common" file_type="Linker Script" />
-      </file>
-      <file file_name="$(DeviceMemoryMap)">
-        <configuration Name="Common" file_type="Memory Map" />
-      </file>
-      <file file_name="../../nRF/Scripts/nRF_Target.js">
-        <configuration Name="Common" file_type="Reset Script" />
-      </file>
-    </folder>
-    <folder Name="Source Files">
-      <configuration Name="Common" filter="c;cpp;cxx;cc;h;s;asm;inc" />
-      <file file_name="01bsp_lighthouse_4.c" />
-    </folder>
-    <folder Name="System Files">
-      <file file_name="$(DeviceCommonVectorsFile)" />
-      <file file_name="$(DeviceVectorsFile)">
-        <configuration Name="Common" file_type="Assembly" />
-      </file>
-      <file file_name="$(SeggerThumbStartup)" />
-    </folder>
-  </project>
   <project Name="01bsp_motors">
     <configuration
       Name="Common"
